requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
requests==2.31.0
pyperclip==1.8.2
PyPDF2==3.0.1
python-docx==1.1.2